chore: skip default cni e2e test on newer k8ss

+++ test/e2e/helm/kuma_helm_deploy_multi_apps.go
@@ -17,13 +17,21 @@ import (
 
 func AppDeploymentWithHelmChart() {
 	var cluster Cluster
+	var skip bool
 
 	minReplicas := 3
 
-	var setup = func(withCni KumaDeploymentOption) {
+	var setup = func(withCni KumaDeploymentOption, shouldSkip func(cluster Cluster) (string, bool)) {
 		cluster = NewK8sCluster(NewTestingT(), Kuma1, Silent).
 			WithTimeout(6 * time.Second).
 			WithRetries(60)
+
+		if shouldSkip != nil {
+			var msg string
+			if msg, skip = shouldSkip(cluster); skip {
+				Skip(msg)
+			}
+		}
 
 		err := NewClusterSetup().
 			Install(Kuma(core.Standalone,
@@ -45,6 +53,10 @@ func AppDeploymentWithHelmChart() {
 	}
 
 	E2EAfterEach(func() {
+		if skip {
+			return
+		}
+
 		Expect(cluster.DeleteNamespace(TestNamespace)).To(Succeed())
 		Expect(cluster.DeleteKuma()).To(Succeed())
 		Expect(cluster.DismissCluster()).To(Succeed())
@@ -52,8 +64,8 @@ func AppDeploymentWithHelmChart() {
 
 	DescribeTable(
 		"Should deploy two apps",
-		func(withCni KumaDeploymentOption) {
-			setup(withCni)
+		func(withCni KumaDeploymentOption, shouldSkip func(cluster Cluster) (string, bool)) {
+			setup(withCni, shouldSkip)
 
 			clientPodName, err := PodNameOfApp(cluster, "demo-client", TestNamespace)
 			Expect(err).ToNot(HaveOccurred())
@@ -79,6 +91,30 @@ func AppDeploymentWithHelmChart() {
 				g.Expect(stderr).To(ContainSubstring("HTTP/1.1 200 OK"))
 			}).Should(Succeed())
 		},
-		Entry("with new cni (experimental)", WithExperimentalCNI()),
+		Entry("with default cni", WithCNI(), func(cluster Cluster) (string, bool) {
+			version, err := cluster.GetK8sVersion()
+			Expect(err).To(Succeed())
+
+			// Just future proofing
+			if version.Major != 1 {
+				return fmt.Sprintf(
+					"default cni is not supported in version %d.%d.%d [supported <= 1.21.x]",
+					version.Major, version.Minor, version.Patch,
+				), true
+			}
+
+			// k3s from version 1.22 comes with flannel CNI plugin in version 1,
+			// which is not supported with our default/legacy kuma-cni plugin
+			// (max supported version is 0.4)
+			if version.Minor > 21 {
+				return fmt.Sprintf(
+					"default cni is not supported in version 1.%d.%d [supported <= 1.21.x]",
+					version.Minor, version.Patch,
+				), true
+			}
+
+			return "", false
+		}),
+		Entry("with new cni (experimental)", WithExperimentalCNI(), nil),
 	)
 }
